requirement.txt
Django==4.1.2
psycopg2-binary==2.9.4
sqlparse==0.4.3
## The following requirements were added by pip freeze:
chardet==5.0.0
defusedxml==0.7.1
diff-match-patch==20200713
django-crispy-forms==1.14.0
django-excel==0.0.10
django-import-export==2.5.0
django-safedelete==1.3.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
lml==0.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyexcel-xlsx==0.6.0
pytz==2022.5
PyYAML==6.0
tablib==3.2.1
texttable==1.6.4
xlrd==2.0.1
xlwt==1.3.0